[FIX] CL warning for equality test between signed and unsigned integer.
git-svn-id: https://b-tk.googlecode.com/svn/BTK/trunk@370 93be6e1e-6fcc-11de-baec-69240fdc299a

+++ Code/BasicFilters/btkForcePlatformsExtractor.cpp
@@ -317,7 +317,7 @@ namespace btk
     {
       pValue = pCalMatrix->GetInfo();
       ForcePlatform::CalMatrix cal = fp->GetCalMatrix();
-      if (pValue->GetValues().size() >= (coefficientsAlreadyExtracted + cal.size()))
+      if (static_cast<int>(pValue->GetValues().size()) >= (coefficientsAlreadyExtracted + cal.size()))
       {
         for (int i = 0 ; i < cal.cols() ; ++i)
           for (int j = 0 ; j < cal.rows() ; ++j)
